* Set Content::m_chunkSize before adding files.
* Removed the old FileList and replaced it with the cleaned up
EntryList code. The API now gives a FileList* instead of FileList.


git-svn-id: svn://rakshasa.no/libtorrent/trunk/libtorrent@810 e378c898-3ddf-0310-93e7-cc216c733640

+++ src/torrent/file.cc
@@ -82,7 +82,7 @@ File::is_correct_size() const {
   if (!fs.update(m_fileMeta->get_path()))
     return false;
 
-  return fs.is_regular() && fs.size() == m_size;
+  return fs.is_regular() && (uint64_t)fs.size() == m_size;
 }
 
 bool
